permissions bug edited and redirection done

+++ application/modules/admin/models/DbTable/Permissions.php
@@ -35,8 +35,14 @@ class Admin_Model_DbTable_Permissions extends Zend_Db_Table_Abstract
     {
         try 
         {
-            $result = $this->delete('role_id='.$roleId." and resource_id=".$resourceId." and action_id=".$actionId);
-            return $return;
+            $count = count($this->getActionId($resourceId, $roleId));
+            
+            if($count !=0)
+            {   
+                $result = $this->delete('role_id='.$roleId." and resource_id=".$resourceId." and action_id=".$actionId);
+                return $result;
+            }
+            
         
         }
         catch(Exception $e)
